+++ lib/msal-common/src/utils/Constants.ts
@@ -49,7 +49,7 @@ export const Constants = {
     KNOWN_PUBLIC_CLOUDS: ["login.microsoftonline.com", "login.windows.net", "login.microsoft.com", "sts.windows.net"],
     TOKEN_RESPONSE_TYPE: "token",
     ID_TOKEN_RESPONSE_TYPE: "id_token",
-    SHR_NONCE_VALIDITY: 240,
+    SHR_NONCE_VALIDITY: 240
 };
 
 export const OIDC_DEFAULT_SCOPES = [
@@ -133,7 +133,7 @@ export enum AADServerParamKeys {
     X_APP_NAME = "x-app-name",
     X_APP_VER = "x-app-ver",
     POST_LOGOUT_URI = "post_logout_redirect_uri",
-    ID_TOKEN_HINT= "id_token_hint",
+    ID_TOKEN_HINT = "id_token_hint",
     DEVICE_CODE = "device_code",
     CLIENT_SECRET = "client_secret",
     CLIENT_ASSERTION = "client_assertion",