+++ samples/msal-node-samples/silent-flow/test/silent-flow-b2c-aad.spec.ts
@@ -4,7 +4,7 @@
  */
 
 import puppeteer from "puppeteer";
-import { Screenshot, createFolder, setupCredentials, b2cAadPpeAccountEnterCredentials } from "../../../e2eTestUtils/TestUtils";
+import { Screenshot, createFolder, setupCredentials, b2cAadPpeAccountEnterCredentials, ONE_SECOND_IN_MS } from "../../../e2eTestUtils/TestUtils";
 import { NodeCacheTestUtils } from "../../../e2eTestUtils/NodeCacheTestUtils";
 import { LabClient } from "../../../e2eTestUtils/LabClient";
 import { LabApiQueryParams } from "../../../e2eTestUtils/LabApiQueryParams";
@@ -33,7 +33,7 @@ const config = require("../config/B2C-AAD.json");
 
 describe("Silent Flow B2C Tests (aad account)", () => {
     jest.retryTimes(1);
-    jest.setTimeout(45000);
+    jest.setTimeout(ONE_SECOND_IN_MS*45);
     let browser: puppeteer.Browser;
     let context: puppeteer.BrowserContext;
     let page: puppeteer.Page;
@@ -88,7 +88,7 @@ describe("Silent Flow B2C Tests (aad account)", () => {
         beforeEach(async () => {
             context = await browser.createIncognitoBrowserContext();
             page = await context.newPage();
-            page.setDefaultTimeout(5000);
+            page.setDefaultTimeout(ONE_SECOND_IN_MS*5);
             await page.goto(homeRoute, { waitUntil: "networkidle0" });
         });
 
@@ -104,7 +104,7 @@ describe("Silent Flow B2C Tests (aad account)", () => {
             await b2cAadPpeAccountEnterCredentials(page, screenshot, username, accountPwd);
             await page.waitForSelector("#acquireTokenSilent");
             await page.click("#acquireTokenSilent");
-            const cachedTokens = await NodeCacheTestUtils.waitForTokens(TEST_CACHE_LOCATION, 2000);
+            const cachedTokens = await NodeCacheTestUtils.waitForTokens(TEST_CACHE_LOCATION, ONE_SECOND_IN_MS*2);
             expect(cachedTokens.accessTokens.length).toBe(1);
             expect(cachedTokens.idTokens.length).toBe(1);
             expect(cachedTokens.refreshTokens.length).toBe(1);
@@ -129,14 +129,17 @@ describe("Silent Flow B2C Tests (aad account)", () => {
             await b2cAadPpeAccountEnterCredentials(page, screenshot, username, accountPwd);
             await page.waitForSelector("#acquireTokenSilent");
 
-            let tokens = await NodeCacheTestUtils.waitForTokens(TEST_CACHE_LOCATION, 2000);
+            let tokens = await NodeCacheTestUtils.waitForTokens(TEST_CACHE_LOCATION, ONE_SECOND_IN_MS*2);
             const originalAccessToken = tokens.accessTokens[0];
             await NodeCacheTestUtils.expireAccessTokens(TEST_CACHE_LOCATION);
-            tokens = await NodeCacheTestUtils.waitForTokens(TEST_CACHE_LOCATION, 2000);
+            tokens = await NodeCacheTestUtils.waitForTokens(TEST_CACHE_LOCATION, ONE_SECOND_IN_MS*2);
             const expiredAccessToken = tokens.accessTokens[0];
+
+            // Wait to ensure new token has new iat
+            await new Promise(r => setTimeout(r, ONE_SECOND_IN_MS));
             await page.click("#acquireTokenSilent");
             await page.waitForSelector('#token-acquired-silently');
-            tokens = await NodeCacheTestUtils.waitForTokens(TEST_CACHE_LOCATION, 2000);
+            tokens = await NodeCacheTestUtils.waitForTokens(TEST_CACHE_LOCATION, ONE_SECOND_IN_MS*2);
             const refreshedAccessToken = tokens.accessTokens[0];
             await screenshot.takeScreenshot(page, "acquireTokenSilentGotTokens");
             const htmlBody = await page.evaluate(() => document.body.innerHTML);